requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
datasets
scikit-learn
matplotlib
seaborn
jupyter
